telegram-bot/requirements.txt
aiogram>=3.4.0
pydantic-settings>=2.0.0
google-generativeai>=0.5.0
httpx>=0.27.0
